+++ src/lib/libcmd/grep.c
@@ -180,9 +180,9 @@ struct State_s				/* program state		*/
 	regmatch_t*	pos;		/* match position pointer	*/
 	int		posnum;		/* number of match positions	*/
 
-	int		after;		/* # lines to list after match	*/
-	int		before;		/* # lines to list before match	*/
-	int		list;		/* list files with hits		*/
+	ssize_t		after;		/* # lines to list after match	*/
+	ssize_t		before;		/* # lines to list before match	*/
+	ssize_t		list;		/* list files with hits		*/
 	regflags_t	options;	/* regex options		*/
 
 	unsigned char	any;		/* if any pattern hit		*/
@@ -669,7 +669,7 @@ grep(char* id, int options, int argc, char** argv, Shbltin_t* context)
 	case 'A':
 		if (opt_info.arg)
 		{
-			state.after = (int)strtol(opt_info.arg, &s, 0);
+			state.after = (ssize_t)strtol(opt_info.arg, &s, 0);
 			if (*s || state.after < 0)
 			{
 	badafter:
@@ -683,7 +683,7 @@ grep(char* id, int options, int argc, char** argv, Shbltin_t* context)
 	case 'B':
 		if (opt_info.arg)
 		{
-			state.before = (int)strtol(opt_info.arg, &s, 0);
+			state.before = (ssize_t)strtol(opt_info.arg, &s, 0);
 			if (*s || state.before < 0)
 			{
 	badbefore:
@@ -697,10 +697,10 @@ grep(char* id, int options, int argc, char** argv, Shbltin_t* context)
 	case 'C':
 		if (opt_info.arg)
 		{
-			state.before = (int)strtol(opt_info.arg, &s, 0);
+			state.before = (ssize_t)strtol(opt_info.arg, &s, 0);
 			if (state.before < 0 || (*s && *s != ','))
 				goto badbefore;
-			state.after = (*s == ',') ? (int)strtol(s + 1, &s, 0) : state.before;
+			state.after = (*s == ',') ? (ssize_t)strtol(s + 1, &s, 0) : state.before;
 			if (*s || state.after < 0)
 				goto badafter;
 		}
@@ -711,7 +711,7 @@ grep(char* id, int options, int argc, char** argv, Shbltin_t* context)
 		state.prefix = opt_info.num;
 		break;
 	case 'L':
-		state.list = (int)(-opt_info.num);
+		state.list = (ssize_t)(-opt_info.num);
 		break;
 	case 'N':
 		h = opt_info.arg;
@@ -739,7 +739,7 @@ grep(char* id, int options, int argc, char** argv, Shbltin_t* context)
 		state.options |= REG_ICASE;
 		break;
 	case 'l':
-		state.list = (int)opt_info.num;
+		state.list = (ssize_t)opt_info.num;
 		break;
 	case 'm':
 		state.label = 1;
